+++ api/src/middleware/authenticateRole.ts
@@ -21,14 +21,33 @@ const authenticateRoles = (roles: RoleName[]) => {
   // Unsure if a returned named function can start with an underscore
   // Returns a requestHandler based on an array of role names
   const _authenticateRoles: RequestHandler = async (req, res, next) => {
+    const { username } = matchedData<{ username: string }>(req);
     const { user } = req;
 
+    // If req.params.username exists and
+    //  authenticated username does not match and
+    //  does not have certain roles throw an error
+    if (username !== user.username) {
+      // User must be an admin
+    }
     // const userHasRole = user.roles.some(
     //   (role) => role.roleDetails.name === roleName
     // );
 
+    const userHasRole = roles.reduce((accum, currentRole) => {
+      return (
+        accum ||
+        user.roles.some((userRole) => userRole.roleDetails.name === currentRole)
+      );
+    }, false);
+
+    console.group("_authenticateRoles...");
+    console.log("userHasRole:", userHasRole);
+    console.groupEnd();
+    if (!userHasRole) {
+    }
+    // If current user does not have roleName throw error
     // if (!userHasRole) throw new InsufficientPermissionsError(roleName);
-    // If current user has roleName throw error
     next();
   };