+++ api/src/services/blog.ts
@@ -1,5 +1,6 @@
 import prisma from "../config/prisma";
-import { PostNotFoundError } from "../errors/customErrors";
+import { PostNotFoundError, UserNotFoundError } from "../errors/customErrors";
+import { Prisma } from "../prisma/generated/prisma";
 
 interface PostContent {
   blogId: string;
@@ -102,14 +103,41 @@ export const getPost = async (publicId: number, username: string) => {
   return post;
 };
 
-export const getPosts = async (blogId: string) => {
-  const posts = await prisma.post.findMany({
+// Should I use prisma.$transaction to get the user, then the user's posts?
+// https://www.prisma.io/docs/orm/prisma-client/queries/transactions#interactive-transactions-1
+export const getPosts = async (username: string) => {
+  /* const posts = await prisma.post.findMany({
     where: {
       blogId: blogId,
     },
   });
 
-  return posts;
+  return posts; */
+  return await prisma.$transaction(async (tx) => {
+    const user = await tx.user
+      .findUniqueOrThrow({
+        where: {
+          username: username,
+        },
+        omit: {
+          password: true,
+        },
+        include: {
+          blog: true,
+        },
+      })
+      .catch(() => {
+        throw new UserNotFoundError(username!, false);
+      });
+
+    const posts = await tx.post.findMany({
+      where: {
+        blogId: user.blog?.id,
+      },
+    });
+
+    return posts;
+  });
 };
 
 export const getPostComment = async ({